requirements.txt
asgiref==3.8.1
attrs==24.3.0
base58==2.1.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.5
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
docker==7.1.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.12.1
idna==3.10
inflection==0.5.1
ipfshttpclient==0.8.0a2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multiaddr==0.0.9
netaddr==1.3.0
psycopg==3.2.4
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
varint==1.0.2
yml==0.0.1
